move to rpc-types
