+++ crates/consensus/Cargo.toml
@@ -15,8 +15,7 @@ exclude.workspace = true
 alloy-primitives = { workspace = true, features = ["rlp"] }
 alloy-rlp.workspace = true
 alloy-eips.workspace = true
-alloy-rpc-types.workspace = true
-thiserror.workspace = true
+thiserror = { workspace = true, optional = true }
 c-kzg = { version = "1.0", features = ["serde"], optional = true }
 sha2 = { version = "0.10", optional = true }
