+++ crates/consensus/src/transaction/eip2930.rs
@@ -4,8 +4,6 @@ use alloy_primitives::{keccak256, Bytes, ChainId, Signature, TxKind, U256};
 use alloy_rlp::{length_of_length, BufMut, Decodable, Encodable, Header};
 use std::mem;
 
-use super::ConversionError;
-
 /// Transaction with an [`AccessList`] ([EIP-2930](https://eips.ethereum.org/EIPS/eip-2930)).
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Default)]
 pub struct TxEip2930 {
@@ -292,26 +290,6 @@ impl Decodable for TxEip2930 {
         }
 
         Self::decode_fields(data)
-    }
-}
-
-impl TryFrom<alloy_rpc_types::Transaction> for Signed<TxEip2930> {
-    type Error = ConversionError;
-
-    fn try_from(tx: alloy_rpc_types::Transaction) -> Result<Self, Self::Error> {
-        let signature = tx.signature.ok_or(ConversionError::MissingSignature)?.try_into()?;
-
-        let tx = TxEip2930 {
-            chain_id: tx.chain_id.ok_or(ConversionError::MissingChainId)?,
-            nonce: tx.nonce,
-            gas_price: tx.gas_price.ok_or(ConversionError::MissingGasPrice)?.to(),
-            gas_limit: tx.gas.to(),
-            to: tx.to.into(),
-            value: tx.value,
-            input: tx.input,
-            access_list: tx.access_list.ok_or(ConversionError::MissingAccessList)?.into(),
-        };
-        Ok(tx.into_signed(signature))
     }
 }
