+++ crates/consensus/src/transaction/envelope.rs
@@ -3,9 +3,6 @@ use crate::{
 };
 use alloy_eips::eip2718::{Decodable2718, Eip2718Error, Encodable2718};
 use alloy_rlp::{Decodable, Encodable, Header};
-use alloy_rpc_types::Transaction;
-
-use super::ConversionError;
 
 /// Ethereum `TransactionType` flags as specified in EIPs [2718], [1559], and
 /// [2930].
@@ -244,22 +241,6 @@ impl Encodable2718 for TxEnvelope {
             TxEnvelope::Eip4844(tx) => {
                 tx.tx().encode_with_signature(tx.signature(), out, false);
             }
-        }
-    }
-}
-
-impl TryFrom<Transaction> for TxEnvelope {
-    type Error = ConversionError;
-
-    fn try_from(tx: Transaction) -> Result<Self, Self::Error> {
-        match tx.transaction_type {
-            None => Ok(Self::Legacy(tx.try_into()?)),
-            Some(t) => match TxType::try_from(t.to::<u8>())? {
-                TxType::Eip1559 => Ok(Self::Eip1559(tx.try_into()?)),
-                TxType::Eip2930 => Ok(Self::Eip2930(tx.try_into()?)),
-                TxType::Eip4844 => Ok(Self::Eip4844(tx.try_into()?)),
-                TxType::Legacy => unreachable!(),
-            },
         }
     }
 }
