+++ crates/consensus/src/transaction/mod.rs
@@ -115,26 +115,3 @@ impl<S: 'static> dyn SignableTransaction<S> {
         }
     }
 }
-
-/// Error variants when parsing a [BlockNumberOrTag]
-#[derive(Debug, thiserror::Error)]
-pub enum ConversionError {
-    #[error("missing `gasPrice` field for Legacy transaction")]
-    MissingGasPrice,
-    #[error("missing signature for transaction")]
-    MissingSignature,
-    #[error("missing `accessList` field for EIP-2930 transaction")]
-    MissingAccessList,
-    #[error("missing `maxFeePerGas` field for EIP-1559 transaction")]
-    MissingMaxFeePerGas,
-    #[error("missing `maxPriorityFeePerGas` field for EIP-1559 transaction")]
-    MissingMaxPriorityFeePerGas,
-    #[error("missing `maxFeePerBlobGas` field for EIP-1559 transaction")]
-    MissingMaxFeePerBlobGas,
-    #[error("missing `chainId` field for EIP-155 transaction")]
-    MissingChainId,
-    #[error(transparent)]
-    SignatureError(#[from] alloy_primitives::SignatureError),
-    #[error(transparent)]
-    Eip2718Error(#[from] alloy_eips::eip2718::Eip2718Error),
-}
