+++ crates/eips/Cargo.toml
@@ -14,7 +14,6 @@ exclude.workspace = true
 [dependencies]
 alloy-primitives = { workspace = true, features = ["rlp"], default-features = false }
 alloy-rlp = { workspace = true, features = ["derive"] }
-alloy-rpc-types.workspace = true
 
 # serde
 serde = { workspace = true, default-features = false, optional = true }
