+++ crates/rpc-types/Cargo.toml
@@ -16,6 +16,8 @@ exclude.workspace = true
 alloy-rlp = { workspace = true, features = ["arrayvec", "derive"] }
 alloy-primitives = { workspace = true, features = ["rlp", "serde", "std"] }
 alloy-serde.workspace = true
+alloy-consensus.workspace = true
+alloy-eips = { workspace = true, features = ["std"] }
 
 ethereum_ssz_derive = { workspace = true, optional = true }
 ethereum_ssz = { workspace = true, optional = true }
